chore: utilize dry principle
